Remove outdated test (#81)
* chore: Update NodeMonitoringConnectionPluginTest
  - Remove outdated test case.
  - Remove NodeMonitoringConnectionPluginBaseTest.java since there is only one test class for
NodeMonitoringConnectionPlugin remaining.
