+++ src/test/java/com/mysql/cj/jdbc/ha/ca/plugins/NodeMonitoringConnectionPluginTest.java
@@ -27,23 +27,34 @@
 package com.mysql.cj.jdbc.ha.ca.plugins;
 
 import com.mysql.cj.conf.DefaultPropertySet;
+import com.mysql.cj.conf.HostInfo;
+import com.mysql.cj.conf.PropertyKey;
 import com.mysql.cj.conf.PropertySet;
+import com.mysql.cj.conf.RuntimeProperty;
+import com.mysql.cj.jdbc.JdbcConnection;
 import com.mysql.cj.jdbc.ha.ca.ClusterAwareConnectionProxy;
 import com.mysql.cj.log.Log;
 import com.mysql.cj.log.NullLogger;
-import org.jboss.invocation.InvocationException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
-import java.util.concurrent.ExecutionException;
+import java.sql.ResultSet;
+import java.sql.Statement;
+import java.util.concurrent.Callable;
+import java.util.function.Supplier;
 import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.anyInt;
+import static org.mockito.ArgumentMatchers.anySet;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.atMostOnce;
 import static org.mockito.Mockito.eq;
@@ -52,10 +63,31 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class NodeMonitoringConnectionPluginTest extends NodeMonitoringConnectionPluginBaseTest {
-  private static final ExecutionException EXECUTION_EXCEPTION = new ExecutionException(
-      "exception",
-      new InvocationException("exception", new Throwable()));
+class NodeMonitoringConnectionPluginTest {
+  @Mock ClusterAwareConnectionProxy proxy;
+  @Mock JdbcConnection connection;
+  @Mock Statement statement;
+  @Mock ResultSet resultSet;
+  @Mock PropertySet propertySet;
+  @Mock HostInfo hostInfo;
+  @Mock IConnectionPlugin mockPlugin;
+  @Mock Log logger;
+  @Mock Supplier<IMonitorService> supplier;
+  @Mock MonitorConnectionContext context;
+  @Mock IMonitorService monitorService;
+  @Mock Callable<?> sqlFunction;
+  @Mock RuntimeProperty<Boolean> failureDetectionEnabledProperty;
+  @Mock RuntimeProperty<Integer> failureDetectionTimeProperty;
+  @Mock RuntimeProperty<Integer> failureDetectionIntervalProperty;
+  @Mock RuntimeProperty<Integer> failureDetectionCountProperty;
+
+  static final String NODE = "node";
+  static final Class<JdbcConnection> MONITOR_METHOD_INVOKE_ON = JdbcConnection.class;
+  static final String MONITOR_METHOD_NAME = "executeQuery";
+  static final String NO_MONITOR_METHOD_NAME = "foo";
+  static final int FAILURE_DETECTION_TIME = 10;
+  static final int FAILURE_DETECTION_INTERVAL = 100;
+  static final int FAILURE_DETECTION_COUNT = 5;
 
   private NodeMonitoringConnectionPlugin plugin;
   private AutoCloseable closeable;
@@ -108,23 +140,6 @@ class NodeMonitoringConnectionPluginTest extends NodeMonitoringConnectionPluginB
     verify(mockPlugin).execute(any(Class.class), eq(NO_MONITOR_METHOD_NAME), eq(sqlFunction));
   }
 
-  @Test
-  void test_4_executeThrowsExecutionException() {
-    when(failureDetectionEnabledProperty.getValue())
-        .thenReturn(Boolean.TRUE);
-
-    initializePlugin();
-
-    assertThrows(Exception.class, () -> {
-      when(mockPlugin.execute(any(Class.class), any(), any()))
-          .thenThrow(EXECUTION_EXCEPTION);
-
-      plugin.execute(MONITOR_METHOD_INVOKE_ON, MONITOR_METHOD_NAME, sqlFunction);
-    });
-
-    verify(monitorService).stopMonitoring(eq(context));
-  }
-
   /**
    * Generate different sets of method arguments where one argument is null to ensure
    * {@link NodeMonitoringConnectionPlugin#NodeMonitoringConnectionPlugin(ICurrentConnectionProvider, PropertySet, IConnectionPlugin, Log)}
@@ -148,5 +163,49 @@ class NodeMonitoringConnectionPluginTest extends NodeMonitoringConnectionPluginB
 
   private void initializePlugin() {
     plugin = new NodeMonitoringConnectionPlugin(proxy, propertySet, mockPlugin, logger, supplier);
+  }
+
+  void initDefaultMockReturns() throws Exception {
+    when(hostInfo.getHost())
+        .thenReturn(NODE);
+    when(supplier.get())
+        .thenReturn(monitorService);
+    when(monitorService.startMonitoring(
+        any(JdbcConnection.class),
+        anySet(),
+        any(HostInfo.class),
+        any(PropertySet.class),
+        anyInt(),
+        anyInt(),
+        anyInt()))
+        .thenReturn(context);
+
+    when(mockPlugin.execute(any(Class.class), anyString(), Mockito.any(Callable.class))).thenReturn("done");
+
+    when(proxy.getCurrentConnection()).thenReturn(connection);
+    when(proxy.getCurrentHostInfo()).thenReturn(hostInfo);
+    when(connection.createStatement()).thenReturn(statement);
+    when(statement.executeQuery(anyString())).thenReturn(resultSet);
+    when(resultSet.next()).thenReturn(false);
+    when(hostInfo.getHost()).thenReturn("host");
+    when(hostInfo.getHost()).thenReturn("port");
+
+    when(propertySet.getBooleanProperty(Mockito.eq(PropertyKey.failureDetectionEnabled)))
+        .thenReturn(failureDetectionEnabledProperty);
+    when(propertySet.getIntegerProperty(Mockito.eq(PropertyKey.failureDetectionTime)))
+        .thenReturn(failureDetectionTimeProperty);
+    when(propertySet.getIntegerProperty(Mockito.eq(PropertyKey.failureDetectionInterval)))
+        .thenReturn(failureDetectionIntervalProperty);
+    when(propertySet.getIntegerProperty(Mockito.eq(PropertyKey.failureDetectionCount)))
+        .thenReturn(failureDetectionCountProperty);
+
+    when(failureDetectionEnabledProperty.getValue())
+        .thenReturn(Boolean.TRUE);
+    when(failureDetectionTimeProperty.getValue())
+        .thenReturn(FAILURE_DETECTION_TIME);
+    when(failureDetectionIntervalProperty.getValue())
+        .thenReturn(FAILURE_DETECTION_INTERVAL);
+    when(failureDetectionCountProperty.getValue())
+        .thenReturn(FAILURE_DETECTION_COUNT);
   }
 }
